AbstractSphericalHarmonics.jl

AbstractSphericalHarmonics.SpinTensor — Type

struct SpinTensor{D,T}

Tensor of rank D with elements of type T, stored separated by spin weights.

Grouping tensor components by their spin weight allows representing tensors via spin-weighted spherical harmonics, and allows calculating covariant derivatives (covariant with respect to the unit sphere).

Note that we use a convention where m^a m̄_a = 2.

AbstractSphericalHarmonics.SpinTensor — Method

Convert Tensor to SpinTensor

AbstractSphericalHarmonics.Tensor — Type

struct Tensor{D,T}

Tensor of rank D with elements of type T, stored component-wise.

The components are stored "normalized", i.e. projected onto the dyad vectors eθ^a and eϕ^a. This removes singularities at the poles by multiplying the ϕ components with respective powers of sin θ. This also means that covariant (index down) and contravariant (index up) indices are represented in the same way.

eθ^a = [1, 0] eθa = [1, 0] eϕ^a = [0, 1/(sin θ)] eϕa = [1, sin θ]

eθ and eϕ are orthogonal, and both have length 1.

gab = diag[1, (sin θ)^2] gab = eθa eϕb + eθb eϕa = ma m̄b + m̄a mb

Example:

Scalar s: stored as is Vector v^a: store [eθ_a v^a, eϕ_a v_a] Tensor t_ab: store t[1,1] = eθ^a eθ^b t_ab t[1,2] = eθ^a eϕ^b t_ab t[2,1] = eϕ^a eθ^b t_ab t[2,2] = eϕ^a eϕ^b t_ab

AbstractSphericalHarmonics.Tensor — Method

Convert SpinTensor to Tensor

AbstractSphericalHarmonics.tensor_gradient! — Method

Calculate gradient

AbstractSphericalHarmonics.tensor_gradient — Method

Calculate gradient